requirements.txt
grpcio-reflection
protobuf
grpcio == 1.20.1
connexion == 1.1.15
python_dateutil == 2.6.0
setuptools >= 21.0.0
